Visual diff of last QCed version of a file to the local version.

Compares the local version of a script with the most recent QCed version. The output will appear in the viewer and only rows where there have been additions, deletions or modifications in the script will be shown.

diffQced(
  .file,
  .side_by_side = TRUE,
  .ignore_white_space = FALSE,
  .display_entire_file = FALSE
)

Arguments

.file

file path from working directory

.side_by_side

Logical. Should diffs be displayed side by side?

.ignore_white_space

Logical. Should white space be ignored?

.display_entire_file

Logical. Should the entire file be displayed?

Examples

with_demoRepo({
  diffQced("script/data-assembly.R")
})
#> Error in with_demoRepo({    diffQced("script/data-assembly.R")}): could not find function "with_demoRepo"