Outstanding QC check by user

This function checks the QC status of files last authored by the given user in a repository. The information provided includes if a file is awaiting QC or not present in the QC log. Users can also filter the results by days since their last edit, to only look at files edited within a specified period of time.

myQC(.within_days = Inf)

Arguments

.within_days

A numeric value indicating how many days back to look for edits. Default is Inf, meaning no time limit is applied.

Value

A list of data frames, each representing a set of files with the same QC status. The data frames include the following columns:

File: The name of the file.

Day(s) since last edit: The number of days since the last modification.

Examples

if (FALSE) { # \dontrun{
# Call myQC for the current user and look at files edited within the last 14 days
result <- myQC(.within_days = 14)

# Print the QC status for files that are awaiting QC
print(result$`Awaiting QC`)
} # }